force TCP connection to not be persistent
Persistent connection can cause some problems when a service receives a
lot (actually, just 10) of connections.

For example, currently xivo-ctid, using cherrypy, use a thread pool of
10 worker threads. It stops processing new TCP connections when 10
connections are already established. When persistent connection are
used, the 10 TCP connections slots can easily be all taken, resulting in
a sort of denial of service.

It would make more sense to implement this in the server part, in
cherrypy, but I did not find yet how to configure cherrypy to disable
persistent connection.

This should be seen as a temporary solution until we implement something
better.