client: add set_token method
Takes a "token" argument, which is a token ID (not a token dictionary).

In the current implementation, it only changes the value of self.token,
i.e. it does not update any session object nor do anything else since
session objects are currently never cached and are created before each
request.