rename token to token_id in BaseClient
There was a clash between xivo-lib-rest-client and xivo-auth-client
for the "token" attribute. xivo-lib-rest-client was using it to store
a "token ID", while xivo-auth-cient is using it to store a "TokenCommand
instance".

So rename token to token_id in BaseClient. This limits the impact on the
API, since even though "token" was not a private attribute, it doesn't
seems to be reference outside of xivo-lib-rest-client, at least in XiVO
code.